LPS-59594 Refactor packages and names
