+++ modules/util/portal-tools-shard-builder/bnd.bnd
@@ -1,5 +1,6 @@
 Bundle-Name: Liferay Portal Tools Shard Builder
 Bundle-SymbolicName: com.liferay.portal.tools.shard.builder
 Bundle-Version: 1.0.0
+Export-Package: com.liferay.portal.tools.shard.builder.exporter
 Include-Resource:\
 	classes
